Update submodule to include microtask type fix
As [#83329] points out, users of the OpenMP WebAssembly port in upstream
LLVM could run into issues with indirect call type mismatches. [#84355]
fixes that problem, so this commit updates the example to use that fix.
It first changes the `llvm-project` submodule to point upstream, since
all of the work is now merged there, and also changes that submodule to
point directly at the commit for [#84355].

[#83329]: https://github.com/llvm/llvm-project/issues/83329
[#84355]: https://github.com/llvm/llvm-project/pull/84355